+++ packages/vaex-hdf5/vaex/hdf5/export.py
@@ -57,7 +57,7 @@ def export_hdf5_v1(dataset, path, column_names=None, byteorder="=", shuffle=Fals
 
         logger.debug("exporting columns(hdf5): %r" % column_names)
         for column_name in column_names:
-            dtype = dataset.dtype(column_name)
+            dtype = dataset.data_type(column_name)
             if column_name in dataset.get_column_names(strings=True):
                 column = dataset.columns[column_name]
                 shape = (N,) + column.shape[1:]
@@ -147,8 +147,8 @@ def export_hdf5(dataset, path, column_names=None, byteorder="=", shuffle=False, 
                 sparse_groups[id(sparse_matrix)].append(column_name)
                 sparse_matrices[id(sparse_matrix)] = sparse_matrix
                 continue
-            dtype = dataset.dtype(column_name)
-            if column_name in dataset.get_column_names(virtual=False, alias=False):
+            dtype = dataset.data_type(column_name)
+            if column_name in dataset.get_column_names(virtual=False):
                 column = dataset.columns[column_name]
                 shape = (N,) + column.shape[1:]
             else: